Move validations to specifications
