--- packages/app/src/cli/validators/extensions/ui.ts
@@ -1,5 +1,0 @@
-import {UIExtension} from '../../models/app/extensions.js'
-
-export async function validateUIExtensions(extensions: UIExtension[]) {
-  return Promise.all(extensions.map((ext) => ext.preDeployValidation()))
-}
